+++ demo/my-components/custom-tag.js
@@ -1,0 +1,9 @@
+import { LitElement, html } from 'https://cdn.skypack.dev/lit-element'
+
+window.customElements.define('custom-tag', class extends LitElement {
+  render () {
+    return html`
+      <p>This is a very simple Web Component</p>
+    `
+  }
+})