+++ index.js
@@ -1,36 +1,10 @@
-const jsdom = require('jsdom')
-const { JSDOM } = jsdom
+const AddWebComponentDefinitions = require('./add-web-component-definitions')
 
-module.exports = function (eleventyConfig, options) {
-  options.path = options.path || function (tag) { return `/js/components/${tag}/${tag}.js` }
-  options.position = options.position || 'beforeend'
-  options.verbose = options.verbose || false
-  options.quiet = options.quiet || false
-
-  eleventyConfig.addTransform('add-web-component-definitions', function (content, outputPath) {
-    if (outputPath.endsWith('.html')) {
-      if (options.verbose) {
-        console.log('[add-web-component-definitions] Examining ', outputPath)
-      }
-      const dom = new JSDOM(content)
-      const body = dom.window.document.body.innerHTML
-      const regex = /<\/(\w+(-\w+)+)>/g
-      const matches = body.matchAll(regex)
-      const tagsSet = new Set()
-      for (const match of matches) {
-        tagsSet.add(match[1])
-      }
-      if (options.verbose) {
-        console.log(`[add-web-component-definitions] Tags found in ${outputPath}:`, tagsSet)
-      }
-      for (const tag of tagsSet) {
-        if (!options.quiet) {
-          console.log('[add-web-component-definitions] Adding definition for tag: ', tag)
-        }
-        dom.window.document.body.insertAdjacentHTML(options.position, `<script type="module" src="${options.path(tag)}"></script>`)
-      }
-      return dom.serialize()
-    }
-    return content
-  })
+module.exports = {
+  configFunction: function (eleventyConfig, options = {}) {
+    eleventyConfig.addTransform(
+      'add-web-component-definitions',
+      AddWebComponentDefinitions.bind(null, options)
+    )
+  }
 }