+++ CHANGELOG.md
@@ -2,6 +2,39 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [1.3.1](https://github.com/jdvivar/eleventy-plugin-add-web-component-definitions/compare/v1.3.0...v1.3.1) (2020-12-26)
+
+## [1.3.0](https://github.com/jdvivar/eleventy-plugin-add-web-component-definitions/compare/v1.1.1...v1.3.0) (2020-12-26)
+
+
+### Features
+
+* add specifiers to configuration object ([91764e1](https://github.com/jdvivar/eleventy-plugin-add-web-component-definitions/commit/91764e145f3f41b3c364c6efbe2cb4f27a62b608))
+* add tests and demo ([df34655](https://github.com/jdvivar/eleventy-plugin-add-web-component-definitions/commit/df346559bbc4746746f31f09a3f4c0de3c2a0414))
+* adding singleScript option ([125bb16](https://github.com/jdvivar/eleventy-plugin-add-web-component-definitions/commit/125bb164ce75d35603c8457eabe2718737424aa1))
+* improving demo ([6c9818f](https://github.com/jdvivar/eleventy-plugin-add-web-component-definitions/commit/6c9818f4d088f8f40380e125f6a9c0811d3913c9))
+* using a throw statement, with tests ([fce57f3](https://github.com/jdvivar/eleventy-plugin-add-web-component-definitions/commit/fce57f3fd1b36719520789ccca594306b8258cfe))
+* using index.js instead, to follow node package standards ([c2dede8](https://github.com/jdvivar/eleventy-plugin-add-web-component-definitions/commit/c2dede8b5151c3e19d8cc95e4999818ca8fbe960))
+
+
+### Bug Fixes
+
+* configure and test options ([270669e](https://github.com/jdvivar/eleventy-plugin-add-web-component-definitions/commit/270669ed9f0adae83c80420d4d9299d9e1bcc116))
+* lint ([91b02e2](https://github.com/jdvivar/eleventy-plugin-add-web-component-definitions/commit/91b02e292d90e0d7df1c8705356d46da44786b2b))
+
+## [1.2.0](https://github.com/jdvivar/eleventy-plugin-add-web-component-definitions/compare/v1.1.1...v1.2.0) (2020-12-25)
+
+
+### Features
+
+* improving demo ([f1e5653](https://github.com/jdvivar/eleventy-plugin-add-web-component-definitions/commit/f1e5653c745d36ac9a2e681c46b036ce22f4e199))
+* using index.js instead, to follow node package standards ([9387f79](https://github.com/jdvivar/eleventy-plugin-add-web-component-definitions/commit/9387f792c630078ca373b398c1cb5be18ef3ef07))
+
+
+### Bug Fixes
+
+* lint ([67d05e3](https://github.com/jdvivar/eleventy-plugin-add-web-component-definitions/commit/67d05e32aab55f842c22ec0329d34d7395869731))
+
 ### [1.1.2](https://github.com/jdvivar/eleventy-plugin-add-web-component-definitions/compare/v1.1.1...v1.1.2) (2020-07-04)
 
 ### [1.1.1](https://github.com/jdvivar/eleventy-plugin-add-web-component-definitions/compare/v1.1.0...v1.1.1) (2020-07-04)